issue/4 Added Adapt.a11y.ariaLevel
